Move two identical methods to util class
