+++ src/main/java/gregtech/client/model/EmissiveOreBakedModel.java
@@ -2,17 +2,14 @@ package gregtech.client.model;
 
 import gregtech.api.unification.ore.StoneType;
 import gregtech.client.utils.BloomEffectUtil;
+import gregtech.client.utils.RenderUtil;
 import gregtech.common.ConfigHolder;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.renderer.block.model.BakedQuad;
 import net.minecraft.client.renderer.block.model.IBakedModel;
-import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
-import net.minecraft.client.renderer.vertex.VertexFormat;
 import net.minecraft.util.BlockRenderLayer;
 import net.minecraft.util.EnumFacing;
 import net.minecraftforge.client.MinecraftForgeClient;
-import net.minecraftforge.client.model.pipeline.UnpackedBakedQuad;
-import net.minecraftforge.fml.client.FMLClientHandler;
 
 import javax.annotation.Nullable;
 import java.util.ArrayList;
@@ -50,30 +47,10 @@ public class EmissiveOreBakedModel extends OreBakedModel {
         if (this.overlayQuads[index] == null) {
             List<BakedQuad> quads = new ArrayList<>(this.overlay.getQuads(null, side, rand));
             for (int i = 0; i < quads.size(); i++) {
-                quads.set(i, transform(quads.get(i)));
+                quads.set(i, RenderUtil.makeEmissive(quads.get(i)));
             }
             return this.overlayQuads[index] = quads;
         }
         return this.overlayQuads[index];
-    }
-
-    private static BakedQuad transform(BakedQuad quad) {
-        if (FMLClientHandler.instance().hasOptifine()) return quad;
-        VertexFormat format = quad.getFormat();
-        if (!format.getElements().contains(DefaultVertexFormats.TEX_2S)) {
-            format = new VertexFormat(quad.getFormat());
-            format.addElement(DefaultVertexFormats.TEX_2S);
-        }
-        UnpackedBakedQuad.Builder builder = new UnpackedBakedQuad.Builder(format) {
-            @Override
-            public void put(int element, float... data) {
-                if (this.getVertexFormat().getElement(element) == DefaultVertexFormats.TEX_2S)
-                    super.put(element, 480.0f / 0xFFFF, 480.0f / 0xFFFF);
-                else super.put(element, data);
-            }
-        };
-        quad.pipe(builder);
-        builder.setApplyDiffuseLighting(false);
-        return builder.build();
     }
 }
